新型コロナコールセンター相談件数: Rename components/cards/TelephoneAdvisoryReportsNumberCard.vue to components/index/CardsReference/TelephoneAdvisoryReportsNumber/Card.vue
